Added a small_vector implementation
I was planning to implement something similar but I found this
on GitHub and I decided to embed it.
https://github.com/gharveymn/small_vector

It seems well written (certainly better than what I would
do), does exactly what we need, and the licence allows this.

As usual this is putting everyting in PLMD namespace

For reference, this is d98388b21b73be55e8c36ee18809e19a6fa82d34