Fixed stretching of switching functions.
Before this, the stretching was implemented assuming
that the value of the switching function in 0 is 1.
If a user was providing (with lepton) a function for which
s(0)!=1, the stretched version still resulted in s(0)=1.

With this fix, the value of the switching function in 0
is preserved. This also allows to have switching functions that
are infinite in the origin (e.g., a Lennard Jones potential).
In this case the stretching exactly corresponds to a truncated-
and-shifted potential.